requirements.txt
requests>=2.31.0
beautifulsoup4>=4.12.0
SQLAlchemy>=2.0.0
onnxruntime>=1.17.0
numpy>=1.24.0
Pillow>=10.0.0
psycopg2-binary>=2.9.0
